feat: firebase 로직 컴포넌트에서 분리 for 강결합
- index.tsx 에서 분리하고 이후 다른 컴포넌트에서도 적용할 예정
